backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
requests>=2.26.0
ccxt>=4.0.0
pandas>=1.3.0
numpy>=1.21.0
ta>=0.10.0
pydantic>=1.8.0
sqlalchemy>=1.4.0
aiosqlite>=0.17.0
asyncpg>=0.24.0
psycopg2-binary>=2.9.0
greenlet>=1.0.0
jinja2>=3.0.0
google-generativeai>=0.3.0
pywebpush>=1.14.0
cryptography>=41.0.0
